+++ tests/test_path_exists_dir.c
@@ -1,0 +1,22 @@
+#include "fsio.h"
+#include "test.h"
+#include <unistd.h>
+
+
+void test_impl()
+{
+  char *file = "./path_exists_dir";
+
+  fsio_mkdir(file, 0777);
+  bool exists = fsio_file_exists(file);
+  rmdir(file);
+
+  assert_true(!exists);
+}
+
+
+int main()
+{
+  test_run(test_impl);
+}
+